requirements.txt
flask==2.3.3
openai==0.27.8
python-dotenv==1.0.0
httpx==0.23.3 